move shell parser into own package
