+++ gpMgmt/go-utils/src/gp_upgrade/shell_parsers/monitor_shell_parser.go
@@ -1,4 +1,4 @@
-package commands
+package shell_parsers
 
 import (
 	"regexp"
