+++ gpMgmt/go-utils/src/gp_upgrade/shell_parsers/shell_parsers_client_suite_test.go
@@ -1,0 +1,13 @@
+package shell_parsers_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestCommands(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Shell Parsers Suite")
+}
